refactor: move some CJS and ESM code analysis to ext/node (#18789)
