+++ cli/npm/mod.rs
@@ -13,6 +13,6 @@ pub use installer::PackageJsonDepsInstaller;
 pub use registry::CliNpmRegistryApi;
 pub use resolution::NpmResolution;
 pub use resolvers::create_npm_fs_resolver;
+pub use resolvers::CliRequireNpmResolver;
 pub use resolvers::NpmPackageResolver;
 pub use resolvers::NpmProcessState;
-pub use resolvers::RequireNpmPackageResolver;
